docs: add example snippet for adding custom schema for yamlls
I struggled for a while to get the YAML language server to pick up my
custom schema as I was providing the same structure as to the JSON
language server. However, as the YAML language server expects a
`{ url = fileMatch }` this was not working. I figured it'd be helpful
for others to include this in the README.

I also believe #9 can be closed now because there's already a
`require('schemastore').yaml.schemas()` function!